fix: split lifetimes in queries and params
Previously it was not possible to return async streams created using
`.iter()` from functions which receive the client as an argument and
create the statement or parameters locally, because they would all
have to share the same lifetime `'a`. Using different lifetimes for the
client, statement and parameters should solve this problem.