fix: remove type parameter `P` in `query_raw`
Previously it was not possible to use the `.iter()` and `.all()` methods
in functions which had to return an `impl Future<Output = ...> + Send`.
In this case the compiler would complain about a type mismatch caused
by a lifetime requirement introduced in the `Item = P` bound in the
`query_raw` declaration. Removing the type parameter `P` and moving its
bounds to `I::Item` seems to solve the problem.